Create compounding_misfit.py
Demonstrate how model misfit at monthly scale can compound when aggregated to annual scale.  This example intentionally chooses an SMB model-AR model pair that is not a great fit to show effect of compounding errors.  The plot generated here was shown as an example in a colloquium talk and should not be taken as a standard illustration